refactor: move GetTransaction to transaction.go :hammer:
